pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genie-space-generator"
version = "1.0.0"
description = "Generate a fully-configured Databricks Genie space for any industry and use case"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Chad Macumber", email = "cm.macumber@gmail.com"}
]
keywords = ["databricks", "genie", "synthetic-data", "metric-views", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = ["mlflow>=2.0"]

[project.urls]
Homepage = "https://github.com/macumberc/genie-space-generator"
Repository = "https://github.com/macumberc/genie-space-generator"

[project.optional-dependencies]
app = [
    "gradio>=4.0",
    "databricks-sdk",
    "databricks-connect",
]
test = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["genie_space_generator*"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
collect_ignore = ["tests/e2e_test.py"]
